requirements.txt
Flask==2.1.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.1
flatbuffers==1.12
gunicorn==20.0.4
Jinja2==3.0.0
mysql-connector-python==8.0.26
mysqlclient==2.0.3
networkx==2.6.3
python-dotenv==0.21.0
requests==2.22.0
requests-oauthlib==1.3.1
six==1.15.0
SQLAlchemy==1.3.12
Flask-MySQLdb
